fruit/requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
bidict==0.22.1
black==24.1.1
blinker==1.7.0
boto3==1.28.63
botocore==1.31.63
cachetools==5.3.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.10.0
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.12.4
fire==0.5.0
Flask==3.0.0
Flask-SocketIO==5.3.6
fonttools==4.43.1
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.40
google-auth==2.23.3
google-auth-oauthlib==1.0.0
gradio_client==0.7.1
grpcio==1.59.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.18.0
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.14.2
opencv-python==4.6.0.66
opencv-python-headless==4.8.0.74
orjson==3.9.10
packaging==23.2
pandas==2.1.1
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
protobuf==4.24.4
psutil==5.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.2
python-multipart==0.0.6
python-socketio==5.11.0
pytz==2023.3.post1
pyvirtualcam==0.11.0
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.0
roboflow==1.1.7
rpds-py==0.13.2
rsa==4.9
s3transfer==0.7.0
sahi==0.11.14
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.0
semantic-version==2.10.0
shapely==2.0.2
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
supervision==0.15.0
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tk==0.1.0
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.1.6
urllib3==1.26.17
uvicorn==0.24.0.post1
websockets==11.0.3
Werkzeug==3.0.0
wsproto==1.2.0
yolov5==7.0.13
zipp==3.17.0
